Final Cleanup for admin page
