+++ frontend/src/components/admin/AdminList/index.js
@@ -2,10 +2,11 @@ import React, {useState, useEffect} from 'react';
 import ReactTable from "react-table";
 import _ from "underscore";
 
-import GameChangerAPI from '../api/gameChanger-service-api';
-import GCButton from "../common/GCButton";
-import { trackEvent } from '../telemetry/Matomo';
-import {styles, TableRow} from './GCAdminStyles';
+import GameChangerAPI from '../../api/gameChanger-service-api';
+import GCButton from "../../common/GCButton";
+import { trackEvent } from '../../telemetry/Matomo';
+import {styles, TableRow} from '../GCAdminStyles';
+import AdminModal from './AdminModal';
 
 const gameChangerAPI = new GameChangerAPI();
 
@@ -17,6 +18,7 @@ export default (props) => {
     // Component Properties
     const {openAdminModal, setPageToView} = props;
     const [gcAdminTableData, setGCAdminTableData] = useState([]);
+    const [showCreateEditAdminModal, setShowCreateEditAdminModal] = useState(false);
     // Component Methods
 
     const getAdminData = async () => {
@@ -58,7 +60,7 @@ export default (props) => {
                         onClick={() => {
                             trackEvent('GAMECHANGER_Admin', "DeleteAdmin", "onClick", row.value);
                             deleteAdminData(row.value).then(() => {
-                                getAdminData(setGCAdminTableData).then(() => {
+                                getAdminData().then(() => {
                                     setPageToView();
                                 });
                             });
@@ -70,24 +72,27 @@ export default (props) => {
         }
     ];
 
-    return (    
-        <div>
-			<div style={{display: 'flex', justifyContent: 'space-between', margin: '10px 80px'}}>
-				<p style={{...styles.sectionHeader, marginLeft: 0, marginTop: 10}}>Gamechanger Admins</p>
-				<GCButton
-					onClick={() => {
-						trackEvent('GAMECHANGER_Admin', "CreateAdmin", "onClick");
-						openAdminModal()
-					}}
-					style={{minWidth: 'unset'}}
-				>Create Admin</GCButton>
-			</div>
-			<ReactTable
-				data={gcAdminTableData}
-				columns={columns}
-				style={{margin: '0 80px 20px 80px', height: 700}}
-				pageSize={10}
-			/>
-		</div>
+    return ( 
+        <>   
+            <div>
+                <div style={{display: 'flex', justifyContent: 'space-between', margin: '10px 80px'}}>
+                    <p style={{...styles.sectionHeader, marginLeft: 0, marginTop: 10}}>Gamechanger Admins</p>
+                    <GCButton
+                        onClick={() => {
+                            trackEvent('GAMECHANGER_Admin', "CreateAdmin", "onClick");
+                            openAdminModal()
+                        }}
+                        style={{minWidth: 'unset'}}
+                    >Create Admin</GCButton>
+                </div>
+                <ReactTable
+                    data={gcAdminTableData}
+                    columns={columns}
+                    style={{margin: '0 80px 20px 80px', height: 700}}
+                    pageSize={10}
+                />
+            </div>
+            <AdminModal />
+        </>
     )
 }
